component for salariesChart
